Add files via upload
This Install comes with all the support that the program needs. Neon Icons in Yellow. Another Version in Green.
Context Menu is easy to deal with. You can switch to explorer from the context menu ( right click ) in the program.
For Copy and Paste to different drives. Sometimes you need to paste using explorer.

If I had to use explorer all the time I would feel like my soul was being drained from me.
That's why I use this Dual File Manager.  It's a lot easier for me to feel comfortable with it.